+++ mod/Core.cs
@@ -12,7 +12,7 @@ namespace RocketRideHUD {
     public class Core : BaseUnityPlugin {
         public const string PluginGUID = "com.github.end-4.rocketRideHud";
         public const string PluginName = "RocketRideHUD";
-        public const string PluginVersion = "1.0.2";
+        public const string PluginVersion = "1.0.3";
 
         public static string workingPath;
         public static string workingDir;