+++ mod/WallJumpWeaponController.cs
@@ -9,6 +9,8 @@ namespace RocketRideHUD {
     public class WallJumpWeaponController : MonoBehaviour {
         public static WallJumpWeaponController Instance { get; private set; }
 
+        public GameObject panel;
+        private Image bgImage;
         public Image icon;
         public TextMeshProUGUI text;
 
@@ -18,40 +20,70 @@ namespace RocketRideHUD {
 
             TMP_FontAsset font = transform.parent.parent.parent.GetComponentInChildren<TextMeshProUGUI>().font;
 
-            GameObject iconO = new GameObject();
-            iconO.name = "WallJumpsIcon";
-            iconO.layer = 5;
-            iconO.transform.SetParent(transform);
-            iconO.transform.localPosition = new Vector3(85, -35, 0);
+            // Create panel
+            panel = new GameObject("WallJumpPanel");
+            panel.layer = 5;
+            RectTransform rect = panel.AddComponent<RectTransform>();
+            panel.transform.SetParent(transform, false);
+            UpdateAlignment(ConfigManager.weaponWallJumpAlignment.value);
+            rect.sizeDelta = new Vector2(46, 25);
+            bgImage = panel.AddComponent<Image>();
+            bgImage.enabled = ConfigManager.weaponWallJumpAlignment.value != WeaponHudAnchor.ShowInside;
+            var sourceImage = transform.parent.parent.GetComponent<Image>();
+            if (sourceImage != null) {
+                bgImage.sprite = sourceImage.sprite;
+                bgImage.type = sourceImage.type;
+                bgImage.pixelsPerUnitMultiplier = sourceImage.pixelsPerUnitMultiplier;
+                bgImage.color = sourceImage.color;
+                bgImage.material = sourceImage.material;
+            }
+            HudOpenEffect hudOpenEffect = panel.AddComponent<HudOpenEffect>();
+
+            // Add icon
+            GameObject iconO = new GameObject {
+                name = "WallJumpsIcon",
+                layer = 5
+            };
+            iconO.transform.SetParent(panel.transform);
+            iconO.transform.localPosition = new Vector3(-9, -0.5f, 0);
             iconO.transform.localRotation = new Quaternion();
-            iconO.transform.localScale = new Vector3(0.2f, 0.2f, 0.2f);
+            iconO.transform.localScale = new Vector3(0.16f, 0.16f, 0.16f);
             icon = iconO.AddComponent<Image>();
             icon.sprite = SpriteLoader.LoadSpriteFromFile(Path.Combine(Core.workingDir, "assets/wall_jump_weapon_hud.png"));
             icon.color = Core.WeaponColor;
 
-            GameObject textO = new GameObject();
-            textO.name = "WallJumpsText";
-            textO.layer = 5;
-            textO.transform.SetParent(transform);
-            textO.transform.localPosition = new Vector3(65, -34, 0);
+            // Add text
+            GameObject textO = new GameObject {
+                name = "WallJumpsText",
+                layer = 5
+            };
+            textO.transform.SetParent(panel.transform);
+            textO.transform.localPosition = new Vector3(8, 0, 0);
             textO.transform.localRotation = new Quaternion();
-            textO.transform.localScale = new Vector3(0.6f, 0.6f, 0.6f);
+            textO.transform.localScale = new Vector3(1f, 1f, 1f);
             text = textO.AddComponent<TextMeshProUGUI>();
             text.font = font;
+            text.fontSize = 18;
             text.alignment = TextAlignmentOptions.Center;
             text.text = Core.MaxWalljumps.ToString();
             text.color = Core.WeaponColor;
 
-            SetStuffActive(ConfigManager.weaponWallJumpShow.value);
+            SetStuffActive(ConfigManager.weaponWallJumpAlignment.value != WeaponHudAnchor.Hidden);
+
+            // Subscribe to events
             NewMovementListener.OnWallJumpsChanged += SetWallJumps;
             PowerUpMeterListener.OnPowerUpStarted += OnPowerUpChange;
             PowerUpMeterListener.OnPowerUpEnded += OnPowerUpChange;
+            SpeedometerListener.OnSpeedometerEnabled += OnSpeedometerEnabled;
+            SpeedometerListener.OnSpeedometerDisabled += OnSpeedometerDisabled;
         }
 
         private void OnDestroy() {
             NewMovementListener.OnWallJumpsChanged -= SetWallJumps;
             PowerUpMeterListener.OnPowerUpStarted -= OnPowerUpChange;
             PowerUpMeterListener.OnPowerUpEnded -= OnPowerUpChange;
+            SpeedometerListener.OnSpeedometerEnabled -= OnSpeedometerEnabled;
+            SpeedometerListener.OnSpeedometerDisabled -= OnSpeedometerDisabled;
         }
 
         public void SetWallJumps(int number) {
@@ -64,6 +96,9 @@ namespace RocketRideHUD {
             SetWallJumps(NewMovement.Instance.gc.onGround ? Core.MaxWalljumps : NewMovement.Instance.currentWallJumps);
         }
 
+        private void OnSpeedometerEnabled() => UpdateSpeedometerAdjustment(true);
+        private void OnSpeedometerDisabled() => UpdateSpeedometerAdjustment(false);
+
         public void UpdateColor() {
             if (icon == null || text == null) return;
 
@@ -72,10 +107,51 @@ namespace RocketRideHUD {
         }
 
         public void SetStuffActive(bool active) {
-            if (icon == null || text == null) return;
+            if (panel == null) return;
 
-            icon.gameObject.SetActive(active);
-            text.gameObject.SetActive(active);
+            panel.SetActive(active);
+        }
+
+        private bool speedometerShown;
+        public void UpdateSpeedometerAdjustment(bool speedometerShown) {
+            this.speedometerShown = speedometerShown;
+            UpdateAlignment();
+        }
+
+        public void UpdateAlignment() {
+            UpdateAlignment(ConfigManager.weaponWallJumpAlignment.value);
+        }
+
+        public void UpdateAlignment(WeaponHudAnchor newValue) {
+            if (newValue == WeaponHudAnchor.Hidden) {
+                SetStuffActive(false);
+            } else {
+                bool rocketRideSameAnchor = newValue == ConfigManager.weaponRocketAlignment.value;
+                RectTransform rect = panel.GetComponent<RectTransform>();
+                switch (newValue) {
+                    case WeaponHudAnchor.ShowTopLeft:
+                        rect.anchoredPosition = new Vector2(rocketRideSameAnchor ? -30 : -77, speedometerShown ? 89 : 63);
+                        break;
+                    case WeaponHudAnchor.ShowTopRight:
+                        rect.anchoredPosition = new Vector2(124, rocketRideSameAnchor ? 89 : 63);
+                        break;
+                    case WeaponHudAnchor.ShowLeft:
+                        rect.anchoredPosition = new Vector2(-124, rocketRideSameAnchor ? 11 : 37);
+                        break;
+                    case WeaponHudAnchor.ShowRight:
+                        rect.anchoredPosition = new Vector2(171, rocketRideSameAnchor ? 11 : 38);
+                        break;
+                    case WeaponHudAnchor.ShowBottom:
+                        rect.anchoredPosition = new Vector2(rocketRideSameAnchor ? -30 : -77, -110);
+                        break;
+                    case WeaponHudAnchor.ShowInside:
+                        rect.anchoredPosition = new Vector2(77, -38);
+                        break;
+                }
+                if (bgImage == null) return;
+                bgImage.enabled = newValue != WeaponHudAnchor.ShowInside;
+                SetStuffActive(true);
+            }
         }
     }
 }